requirements.txt
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
annotated-types==0.7.0
anyio==4.4.0
bcrypt==4.2.0
boto3==1.34.153
botocore==1.34.153
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==43.0.0
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.112.0
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
greenlet==3.0.3
h11==0.14.0
idna==3.7
iniconfig==2.0.0
jmespath==1.0.1
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
packaging==24.1
passlib==1.7.4
pluggy==1.5.0
pyasn1==0.6.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.4.0
pydantic_core==2.20.1
PyJWT==2.9.0
pynamodb==6.0.1
pytest==8.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.37.2
tomli==2.0.1
typing_extensions==4.12.2
urllib3==1.26.19
uvicorn==0.30.5
